+++ packages/core/android/src/main/java/io/sentry/react/RNSentryReactFragmentLifecycleTracer.java
@@ -16,14 +16,22 @@ import io.sentry.ILogger;
 import io.sentry.SentryLevel;
 import io.sentry.android.core.BuildInfoProvider;
 import io.sentry.android.core.internal.util.FirstDrawDoneListener;
+import java.util.HashMap;
+import java.util.Map;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 public class RNSentryReactFragmentLifecycleTracer extends FragmentLifecycleCallbacks {
+  /**
+   * The name react-native-screens uses for its screen appear event. Matching on the class name
+   * instead would break, because R8 can merge several event classes into one.
+   */
+  private static final String SCREEN_APPEAR_EVENT_NAME = "topAppear";
 
   private @NotNull final BuildInfoProvider buildInfoProvider;
   private @NotNull final Runnable emitNewFrameEvent;
   private @NotNull final ILogger logger;
+  private final Map<Fragment, EventDispatcherListenerWrapper> listenerWrapperMap = new HashMap<>();
 
   public RNSentryReactFragmentLifecycleTracer(
       @NotNull BuildInfoProvider buildInfoProvider,
@@ -83,18 +91,39 @@ public class RNSentryReactFragmentLifecycleTracer extends FragmentLifecycleCallb
       return;
     }
 
-    final @NotNull Runnable emitNewFrameEvent = this.emitNewFrameEvent;
-    eventDispatcher.addListener(
-        new EventDispatcherListener() {
+    EventDispatcherListenerWrapper listenerWrapper =
+        new EventDispatcherListenerWrapper(eventDispatcher) {
           @Override
           public void onEventDispatch(Event event) {
-            if ("com.swmansion.rnscreens.events.ScreenAppearEvent"
-                .equals(event.getClass().getCanonicalName())) {
-              eventDispatcher.removeListener(this);
+            if (SCREEN_APPEAR_EVENT_NAME.equals(event.getEventName())) {
+              this.dispatcher.removeListener(this);
+              listenerWrapperMap.remove(f);
               FirstDrawDoneListener.registerForNextDraw(v, emitNewFrameEvent, buildInfoProvider);
             }
           }
-        });
+        };
+
+    eventDispatcher.addListener(listenerWrapper);
+    listenerWrapperMap.put(f, listenerWrapper);
+  }
+
+  @Override
+  public void onFragmentViewDestroyed(@NonNull FragmentManager fm, @NonNull Fragment f) {
+    super.onFragmentViewDestroyed(fm, f);
+    EventDispatcherListenerWrapper listenerWrapper = listenerWrapperMap.get(f);
+    if (listenerWrapper != null && listenerWrapper.dispatcher != null) {
+      listenerWrapper.dispatcher.removeListener(listenerWrapper);
+      listenerWrapperMap.remove(f);
+    }
+  }
+
+  abstract static class EventDispatcherListenerWrapper implements EventDispatcherListener {
+
+    protected final EventDispatcher dispatcher;
+
+    public EventDispatcherListenerWrapper(EventDispatcher dispatcher) {
+      this.dispatcher = dispatcher;
+    }
   }
 
   private static @Nullable EventDispatcher getEventDispatcherForReactTag(